+++ sync-for-commerce/src/codatsynccommerce/models/shared/localization.py
@@ -13,7 +13,7 @@ from typing import Optional
 class Localization:
     required: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('required'), 'exclude': lambda f: f is None }})
     r"""`True` if required."""
-    text: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('text'), 'exclude': lambda f: f is None }})
+    text: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('text') }})
     r"""Value of the property."""