+++ sync-for-commerce/RELEASES.md
@@ -569,3 +569,13 @@ Based on:
 - [python v2.2.1] sync-for-commerce
 ### Releases
 - [PyPI v2.2.1] https://pypi.org/project/codat-sync-for-commerce/2.2.1 - sync-for-commerce
+
+## 2023-09-29 00:03:56
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.3.0] sync-for-commerce
+### Releases
+- [PyPI v2.3.0] https://pypi.org/project/codat-sync-for-commerce/2.3.0 - sync-for-commerce